+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -1179,4 +1179,9 @@ it "must have test for CVE-2016-2097" do
   expect(sc).not_to   be_nil
   expect(sc.class).to eq(Dawn::Kb::CVE_2016_2097)
 end
+it "must have test for OSVDB_132234" do
+  sc = kb.find("OSVDB_132234")
+  expect(sc).not_to   be_nil
+  expect(sc.class).to eq(Dawn::Kb::OSVDB_132234)
+end
 end