+++ lib/dawn/core.rb
@@ -105,7 +105,7 @@ module Dawn
     end
 
     def self.find_conf(create_if_none = false)
-      conf_name  = 'dawnscanner.yaml'
+      conf_name  = 'dawnscanner.yml'
       path_order = [
         './',
         '~/',