+++ lib/dawn/kb/dependency_check.rb
@@ -42,10 +42,6 @@ module Dawn
           @mitigated = true if dep[:name] == @aux_mitigation_gem[:name] unless @aux_mitigation_gem.nil?
 
           @safe_dependencies.each do |safe_dep|
-            if safe_dep[:name] == "rails"
-              debug_me "Forcing save_minor flag for rails gem dependency check"
-              self.save_minor = true
-            end
 
             if dep[:name] == safe_dep[:name]
               v = Dawn::Kb::VersionCheck.new(