+++ lib/dawn/knowledge_base.rb
@@ -267,6 +267,10 @@ require "dawn/kb/cve_2016_0752"
 require "dawn/kb/cve_2016_0753"
 require "dawn/kb/cve_2016_2097"
 require "dawn/kb/cve_2016_2098"
+require "dawn/kb/cve_2016_5697"
+require "dawn/kb/cve_2016_6316"
+require "dawn/kb/cve_2016_6317"
+require "dawn/kb/cve_2016_6582"
 
 # OSVDB
 
@@ -286,6 +290,7 @@ require "dawn/kb/osvdb_119927"
 require "dawn/kb/osvdb_120415"
 require "dawn/kb/osvdb_120857"
 require "dawn/kb/osvdb_121701"
+require "dawn/kb/osvdb_132234"
 
 
 
@@ -569,6 +574,10 @@ module Dawn
           Dawn::Kb::CVE_2016_0753.new,
           Dawn::Kb::CVE_2016_2097.new,
           Dawn::Kb::CVE_2016_2098.new,
+          Dawn::Kb::CVE_2016_5697.new,
+          Dawn::Kb::CVE_2016_6316.new,
+          Dawn::Kb::CVE_2016_6317.new,
+          Dawn::Kb::CVE_2016_6582.new,
 
 
           # OSVDB Checks are still here since are all about dependencies
@@ -588,6 +597,7 @@ module Dawn
           Dawn::Kb::OSVDB_120415.new,
           Dawn::Kb::OSVDB_120857.new,
           Dawn::Kb::OSVDB_121701.new,
+          Dawn::Kb::OSVDB_132234.new,
       ]
         # END @cve_security_checks array
         # START @owasp_ror_cheatsheet_checks array
@@ -611,7 +621,7 @@ module Dawn
 
           ret = []
           ret += @aux_checks
-          ret += @cve_security_checks         if @enabled_checks.include?(:cve_bulletin)
+          ret += @cve_security_checks         if @enabled_checks.include?(:bulletin)
           ret += @owasp_ror_cheatsheet_checks if @enabled_checks.include?(:owasp_ror_cheatsheet)
           ret += @code_quality_checks         if @enabled_checks.include?(:code_quality)